Fix fdwait/copy_cgc_fd_set
This bug caused `fdwait` to always return `CGC_EBADF`, and with many of the challenges this would cause them to exit immediately.

The problem was that the fd was being checked to see if it's in the valid range of fds before first checking that this fd was set. This caused `copy_cgc_fd_set` to always return `CGC_EBADF` because `kExpectedMaxFDs < CGC__NFDBITS`
Also a couple invalid dereferences fixed