Force cb-replay to wait until server is ready
Basically the problem was a race condition where the replay would start before the binaries were all running, causing it to immediately fail when it tries to recv anything. Now it waits for a "ready byte" from the server that will be sent right after all the binaries are started.