Make cgc_initialize_flag_page a constructor.
This ensures that the flag page is initialized before `main` starts, and
requires less patching of the CBs. As an additional advantage, the CBs
now play nicer with tools such as AFL.

I also renamed the page from "secret page" or "magic page" to "flag
page"; I believe this is the official CGC terminology.