CMakeLists.txt: Fix incorrect _FORTIFY_SOURCE macro
The fortify source mitigation is set by the _FORTIFY_SOURCE macro set along with the same optimization level. We used a `-DFORTIFY_SOURCE=0` which is invalid and should be `-D_FORTIFY_SOURCE=0` instead.

To be honest, we could even remove this flag as well, but I think it is fine to leave it, but it should be spelled correctly then :).